html/dom/elements/global-attributes/the-translate-attribute/the-translate-attribute-009.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>translate=no</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/dom.html#the-translate-attribute'>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" type="text/css" href="../../../../../testharness.css">
<meta name="assert" content="If the translate attribute is set to no, javascript will detect the translation mode of text as no-translate.">
</head>
<body>



<div class="test"><div id="box" translate="no">&nbsp;</div></div>


<script>
test(function() {
assert_false(document.getElementById('box').translate);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
